Fix clippy warnings

+++ prusti-viper/src/encoder/encoder.rs
@@ -668,7 +668,7 @@ impl<'v, 'tcx> Encoder<'v, 'tcx> {
             ty::TyKind::Uint(ty::UintTy::U16) => (value as u16).into(),
             ty::TyKind::Uint(ty::UintTy::U32) => (value as u32).into(),
             ty::TyKind::Uint(ty::UintTy::U64) => (value as u64).into(),
-            ty::TyKind::Uint(ty::UintTy::U128) => (value as u128).into(),
+            ty::TyKind::Uint(ty::UintTy::U128) => value.into(),
             ty::TyKind::Uint(ty::UintTy::Usize) => (value as usize).into(),
             ty::TyKind::Char => value.into(),
             ref x => unimplemented!("{:?}", x),
